Make option-value components own their styles
