+++ packages/harpguru-core/src/components/option-new/option.tsx
@@ -56,75 +56,39 @@ export const Option = (props: OptionProps): React.ReactElement => {
   const { [innerActiveIdPos - 2]: inactiveOptionId4 } = extendedList
   const { [innerActiveIdPos - 3]: inactiveOptionId5 } = extendedList
 
-  const optionValuesActiveStyle = {
-    ...styles.optionValues,
-    ...styles.optionValuesActive,
-  }
-  const optionValuesStyle = { ...styles.optionValues }
-
   return (
     <View style={[styles.option]}>
       <OptionTitle>{title}</OptionTitle>
-      <View style={styles.optionValues}>
+      <View>
         <OptionValue
           id={inactiveOptionId1}
           isActive={preactiveId === inactiveOptionId1}
           setFunction={setActiveOptionCallback}
-          style={
-            preactiveId === inactiveOptionId1
-              ? optionValuesActiveStyle
-              : optionValuesStyle
-          }
         />
         <OptionValue
           id={inactiveOptionId2}
           isActive={preactiveId === inactiveOptionId2}
           setFunction={setActiveOptionCallback}
-          style={
-            preactiveId === inactiveOptionId2
-              ? optionValuesActiveStyle
-              : optionValuesStyle
-          }
         />
         <OptionValue
           id={activeOptionId}
           isActive={preactiveId === activeOptionId}
           setFunction={setActiveOptionCallback}
-          style={
-            preactiveId === activeOptionId
-              ? optionValuesActiveStyle
-              : optionValuesStyle
-          }
         />
         <OptionValue
           id={inactiveOptionId3}
           isActive={preactiveId === inactiveOptionId3}
           setFunction={setActiveOptionCallback}
-          style={
-            preactiveId === inactiveOptionId3
-              ? optionValuesActiveStyle
-              : optionValuesStyle
-          }
         />
         <OptionValue
           id={inactiveOptionId4}
           isActive={preactiveId === inactiveOptionId4}
           setFunction={setActiveOptionCallback}
-          style={
-            preactiveId === inactiveOptionId4
-              ? optionValuesActiveStyle
-              : optionValuesStyle
-          }
         />
         <OptionValue
           id={inactiveOptionId5}
           isActive={preactiveId === inactiveOptionId5}
           setFunction={setActiveOptionCallback}
-          style={
-            preactiveId === inactiveOptionId5
-              ? optionValuesActiveStyle
-              : optionValuesStyle
-          }
         />
       </View>
     </View>
